# performance_logging_macros