--- a/A3-Antistasi/Includes/common.inc
+++ b/A3-Antistasi/Includes/common.inc
@@ -1,2 +1,3 @@
 //includes commonly included files to clean up includes in other files
 #include "LogMacros.inc"
+#include "performance.inc"