--- a/A3-Antistasi/Includes/performance.inc
+++ b/A3-Antistasi/Includes/performance.inc
@@ -11,7 +11,7 @@
             markBatch(X,C)  - Marks elapsed time with note C
             stopBatch(X)    - Stios counting and logs the marks in the batch
 
-        Note: all arguments for macros MUST be strings
+        Note: X must be none string, C must be string
 
         How to use:
         Before this is included define diag_performance as 1
@@ -27,16 +27,16 @@
 
     Example:
 
-        startCounter("Test");
+        startCounter(Test);
         //some block of code run
         stopCounter("Test"); //logs time code block used
 
-        startBatch("Test");
+        startBatch(Test);
         //some code
-        markBatch("Test", "some done"); //note <<some done: {elapsed time}>> added to batch
+        markBatch(Test, "some done"); //note <<some done: {elapsed time}>> added to batch
         //more code
-        markBatch("Test", "more code done"); //note <<more code done: {elapsed time}>> added to batch
-        stopBatch("Test"); // batch is logged neatly
+        markBatch(Test, "more code done"); //note <<more code done: {elapsed time}>> added to batch
+        stopBatch(Test, "Last comment"); // batch is logged neatly
 
     License: MIT License
 */