## What type of PR is this.
1. [ ] Bug
2. [ ] Change
3. [x] Enhancement

### What have you changed and why?
Information:
    added macros to log performance, this is meant for when your writing your code to test how it performance so you can make adjustments when necessary if something is badly hurting the performance of your code.

# Added macros:
For simple timespan, meant for chucks of code or simple this took x long to run.
- startCounter(x);
- stopCounter(x);

For more complex stuff, like a whole file, where you mark a batch with comments to give you a better idea of how long x,y,z took to do in the larger picture
- startBatch(x);
- markBatch(x,comment);
- stopBatch(x);

### Please specify which Issue this PR Resolves.
No issue, this is simply an extra tool to help the devs during development.

### Please verify the following and ensure all checks are completed.

1. [x] Have you loaded the mission in LAN host?
2. [ ] Have you loaded the mission on a dedicated server?

### Is further testing or are further changes required?
1. [x] No
2. [ ] Yes (Please provide further detail below.)

### How can the changes be tested?
Steps: 
define diag_performance  as 1 then include the performance.inc file.
example for any file under one of the many folders in %root%\functions
```sqf
#define diag_performance 1
#include "..\..\common.inc"
```
then use the macros:
- startCounter(x);
- stopCounter(x);
- startBatch(x);
- markBatch(x,comment);
- stopBatch(x);

to log performance.
```sqf
#define diag_performance 1
#include "..\..\common.inc"
startBatch("ExampleBatch");

//some code block
markBatch("ExampleBatch","some code block done");

//another code block
markBatch("ExampleBatch","another code block done");

stopBatch("ExampleBatch");
```
```sqf
#define diag_performance 1
#include "..\..\common.inc"
startCounter("Example");
//some code block
stopCounter("Example");
```

---
Example log of ``saveLoop.sqf``
```
15:20:36 Performance Counter | Save | Time: 543.945 ms
15:20:36 Performance Batch | Save | [
    Players saved: 0 ms
    Save List retrieved: 0 ms
    Params: 1.95313 ms
    Resource HR Faction garage: 1.95313 ms
    Statics: 2.92969 ms
    Arsenal: 2.92969 ms
    Prestige: 3.90625 ms
    Garrison: 5.85938 ms
    Mines: 5.85938 ms
    FIA outposts: 5.85938 ms
    Tasks: 5.85938 ms
    Idle bases and assets: 5.85938 ms
    killzones: 5.85938 ms
    rebel controls: 6.83594 ms
    ProfileNamespace saved: 543.945 ms
    Batch stopped: 543.945 ms
]
```
********************************************************
Notes: